Top screen is finished + profit lable is changing every 3 min

+++ TradeApp/TradeApp/Main Screen/Top Screen/Trader.swift
@@ -11,6 +11,7 @@ struct Trader {
     
     let country: UIImage?
     let name: String
-    var deposit: Double
-    var profit: Double
+    var deposit: Int
+    var profit: Int
+    
 }
